expand use of enum
