+++ lib/iris/fileformats/nimrod.py
@@ -15,6 +15,7 @@ import numpy as np
 import iris
 from iris.exceptions import TranslationError
 import iris.fileformats.nimrod_load_rules
+from iris.fileformats.nimrod_load_rules import Table
 
 # general header (int16) elements 1-31 (Fortran bytes 1-62)
 general_header_int16s = (
@@ -203,13 +204,6 @@ def _read_chars(infile, num):
     return result
 
 
-class Table(Enum):
-    table_1 = "Table_1"
-    table_2 = "Table_2"
-    table_3 = "Table_3"
-    table_4 = "Table_4"
-
-
 class NimrodField:
     """A data field from a NIMROD file.
 
@@ -303,17 +297,18 @@ class NimrodField:
             self.threshold_value != default_float_threshold
             or self.threshold_value_alt != default_float_threshold
         )
+        # The ENUM class is in iris.fileformats.nimrod_load_rules
         if self.field_code in Table_3_field_codes:
-            table = Table.table_3.value
+            table = Table.table_3
             data_header_int16s = table_3_data_header_int16s
         elif self.field_code in Table_4_field_codes:
-            table = Table.table_4.value
+            table = Table.table_4
             data_header_int16s = table_4_data_header_int16s
         elif threshold_set:
-            table = Table.table_2.value
+            table = Table.table_2
             data_header_int16s = table_2_data_header_int16s
         else:
-            table = Table.table_1.value
+            table = Table.table_1
             data_header_int16s = table_1_data_header_int16s
 
         self.table = table
